# docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.ArticleCreate:
    properties:
      category_id:
        type: string
      content:
        type: string
      cover_image:
        type: string
      excerpt:
        type: string
      published_at:
        type: string
      slug:
        type: string
      status:
        type: string
      tag_ids:
        items:
          type: string
        type: array
      title:
        type: string
      website_id:
        type: string
    required:
    - content
    - slug
    - title
    - website_id
    type: object
  entity.ArticleUpdate:
    properties:
      category_id:
        type: string
      content:
        type: string
      cover_image:
        type: string
      excerpt:
        type: string
      slug:
        type: string
      status:
        type: string
      tag_ids:
        items:
          type: string
        type: array
      title:
        type: string
      website_id:
        type: string
    type: object
...
  response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@news-api.dev
    name: News API Support
  description: REST API untuk website berita. Mendukung auth JWT, artikel, kategori,
    tag, dan komentar.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: News API
  version: "1.0"
paths:
  /articles:
    get:
      description: Menampilkan daftar artikel dengan filter & pagination
      parameters:
      - default: 1
        description: Nomor halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Jumlah per halaman
        in: query
        name: limit
        type: integer
      - description: 'Filter status: draft/published/archived'
        in: query
        name: status
        type: string
      - description: Filter kategori (UUID)
        in: query
        name: category_id
        type: string
      - description: Filter website (UUID)
        in: query
        name: website_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Daftar artikel dengan meta pagination
          schema:
            $ref: '#/definitions/response.Response'
      summary: List artikel
      tags:
      - Articles
    post:
      consumes:
      - application/json
      description: 'Membuat artikel baru (role: author/editor/admin)'
      parameters:
      - description: Data artikel
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.ArticleCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Artikel berhasil dibuat
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Request tidak valid
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Buat artikel
      tags:
      - Articles
  /articles/{id}:
    delete:
      description: 'Menghapus artikel (role: author/editor/admin)'
      parameters:
      - description: ID artikel
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Artikel berhasil dihapus
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Artikel tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus artikel
      tags:
      - Articles
    get:
      description: Mengambil detail artikel berdasarkan ID
      parameters:
      - description: ID artikel
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data artikel
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Artikel tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      summary: Detail artikel by ID
      tags:
      - Articles
    put:
      consumes:
      - application/json
      description: 'Mengubah data artikel (role: author/editor/admin)'
      parameters:
      - description: ID artikel
        in: path
        name: id
        required: true
        type: string
      - description: Data yang diubah
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.ArticleUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Artikel berhasil diubah
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Artikel tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update artikel
      tags:
      - Articles
...
  /articles/{id}/status:
    patch:
      consumes:
      - application/json
      description: 'Mengubah status artikel: draft/published/archived'
      parameters:
      - description: ID artikel
        in: path
        name: id
        required: true
        type: string
      - description: Status baru
        in: body
        name: body
        required: true
        schema:
          properties:
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Status berhasil diubah
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Status tidak valid
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update status artikel
      tags:
      - Articles
  /articles/published:
    get:
      description: Menampilkan artikel yang sudah published (untuk publik), bisa difilter
        per website
      parameters:
      - default: 1
        description: Nomor halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Jumlah per halaman
        in: query
        name: limit
        type: integer
      - description: Filter website (UUID)
        in: query
        name: website_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Daftar artikel published
          schema:
            $ref: '#/definitions/response.Response'
      summary: List artikel published
      tags:
      - Articles
  /articles/slug/{slug}:
    get:
      description: Mengambil detail artikel berdasarkan slug, view count otomatis
        bertambah
      parameters:
      - description: Slug artikel
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data artikel
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Artikel tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      summary: Detail artikel by slug
      tags:
      - Articles
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"